[fix] tests
