+++ frappe/tests/test_domainification.py
@@ -7,6 +7,8 @@ from frappe.core.page.permission_manager.permission_manager import get_roles_and
 from frappe.desk.doctype.desktop_icon.desktop_icon import (get_desktop_icons, add_user_icon,
 	clear_desktop_icons_cache)
 
+from frappe.core.doctype.domain_settings.domain_settings import get_active_modules
+
 class TestDomainification(unittest.TestCase):
 	def setUp(self):
 		# create test domain
@@ -33,7 +35,7 @@ class TestDomainification(unittest.TestCase):
 
 	def remove_from_active_domains(self, domain=None, remove_all=False):
 		""" remove domain from domain settings """
-		if not domain:
+		if not (domain or remove_all):
 			return
 
 		domain_settings = frappe.get_doc("Domain Settings", "Domain Settings")
@@ -112,7 +114,7 @@ class TestDomainification(unittest.TestCase):
 
 	def test_desktop_icon_for_domainification(self):
 		""" desktop icon should be hidden if doctype's restrict to domain is not in active domains """
-		
+
 		test_doctype = self.new_doctype("Test Domainification")
 		test_doctype.restrict_to_domain = "_Test Domain 2"
 		test_doctype.insert()
@@ -144,10 +146,14 @@ class TestDomainification(unittest.TestCase):
 
 		self.add_active_domain("_Test Domain 2")
 
-		modules = frappe.get_active_modules()
+		modules = get_active_modules()
 		self.assertTrue("Contacts" in modules)
 
 		# doctype should be hidden from the desk
 		self.remove_from_active_domains("_Test Domain 2")
-		modules = frappe.get_active_modules()
-		self.assertTrue("Test Module" not in modules)
+		modules = get_active_modules()
+		self.assertTrue("Contacts" not in modules)
+
+		test_module_def = frappe.get_doc("Module Def", "Contacts")
+		test_module_def.restrict_to_domain = ""
+		test_module_def.save()
